+++ Common/src/main/java/fuzs/puzzleslib/api/biome/v1/BiomeModificationContext.java
@@ -1,0 +1,5 @@
+package fuzs.puzzleslib.api.biome.v1;
+
+public record BiomeModificationContext(ClimateSettingsContext climateSettings, SpecialEffectsContext specialEffects, GenerationSettingsContext generationSettings, MobSpawnSettingsContext mobSpawnSettings) {
+
+}