+++ Common/src/main/java/fuzs/puzzleslib/impl/PuzzlesLib.java
@@ -1,10 +1,15 @@
 package fuzs.puzzleslib.impl;
 
+import fuzs.puzzleslib.api.biome.v1.BiomeLoadingPhase;
 import fuzs.puzzleslib.api.networking.v3.NetworkHandler;
 import fuzs.puzzleslib.core.ModConstructor;
 import fuzs.puzzleslib.impl.capability.ClientboundSyncCapabilityMessage;
 import fuzs.puzzleslib.impl.entity.ClientboundAddEntityDataMessage;
 import net.minecraft.network.protocol.game.ClientboundAddEntityPacket;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.MobCategory;
+import net.minecraft.world.level.biome.Biomes;
+import net.minecraft.world.level.biome.MobSpawnSettings;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -20,4 +25,13 @@ public class PuzzlesLib implements ModConstructor {
             .registerClientbound(ClientboundSyncCapabilityMessage.class)
             .registerClientbound(ClientboundAddEntityDataMessage.class)
             .build();
+
+    @Override
+    public void onRegisterBiomeModifications(BiomeModificationsContext context) {
+        context.register(BiomeLoadingPhase.MODIFICATIONS, biomeLoadingContext -> biomeLoadingContext.getResourceKey() == Biomes.PLAINS, biomeModificationContext -> {
+            biomeModificationContext.mobSpawnSettings().clearSpawns(MobCategory.CREATURE);
+            biomeModificationContext.mobSpawnSettings().clearSpawns(MobCategory.MONSTER);
+            biomeModificationContext.mobSpawnSettings().addSpawn(MobCategory.MONSTER, new MobSpawnSettings.SpawnerData(EntityType.ENDERMAN, 10, 4, 4));
+        });
+    }
 }